Reorganization of Files
Added __str__ dunder method to blobs to exist

Fixed Dr Blob's pills again (hopefully)... Boost Pill should now work even if the thing is on cooldown

Added Dr. Blob to the almanac

Separated menus into different files so its easier to read them

Removed extraneous imports from display_graphics
